pyproject.toml
[tool.poetry]
name = "nanonispy"
version = "1.0.6"
description = "Read Nanonis data files"
authors = ["Yann-Sebastien Tremblay-Johnston <yanns.tremblay@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.17"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.8"
pytest = "^5.3"
coverage = "^4.5"
pytest-sugar = "^0.9.2"
pylint = "^2.4.4"
black = {version = "^19.10b0", allow-prereleases = true}
python-dateutil = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
